% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg2env}
\alias{pkg2env}
\alias{pkg2env_}
\title{Get environment of a package.}
\usage{
pkg2env(pkg)

pkg2env_(pkg)
}
\arguments{
\item{pkg}{A package.}
}
\value{
the environment corresponding to \code{pkg}.
}
\description{
Utility function to get the environment of a package on the search 
path.
}
\examples{
# Non-standard evaluation version
pkg2env(graphics)

# Standard evaluations versions
pkg2env_("tools")
pkg2env_(~ utils)
pkg2env_(quote(stats))
}
\seealso{
\code{\link[base]{list2env}}
}
